Create new TCAD test for Philips model
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/23

The new implementation of the Philips moblity model contains a Newton
loop to find the minimum value for G(P).  This loop is only executed if
the temperature is a non-default value.   Most TCAD device tests in the
test suite use default temperature.

This commit adds a test that is:

1. otherwise identical to the npn_bjt_pde.cir test.
2. sets the temperature to 35C  (default is 27).
3. uses the Philips model, which is the whole point.