Fix bug_1558 test script
Since it was first created in 2009 with commit  c251ce, the
bug_1558.cir.sh script has improperly referenced nonexistent files.

it tries to diff output against 'subckt_0' files that clearly call
back to a SUBCKT test case from which the script was created.

Fortunately, if diff fails then it falls back on doing xyce_verify,
and that has always passed.  But diff has *always* failed because it's
trying to diff against a file that doesn't exist.

Yeesh.