Remove hardcoded library path from tests
The original form of the REST API tests in Matlab
used a hardcoded path the underlying XyceCInterface library.
This isn't needed if the xyce_interface.py file is set
up for CMake to write the library path into it at install
time.  That's been done on the source tree so now
the tests don't need the path.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/477