Add math.h include file
This test failed to compile and GCC appeared to state that
the function fabs() was not defined and that one needed to
include math.h.  A similar test ADCtest.c does include
math.h and it works.  So this was likely a change in how
gcc handles implicitly defined functions.