First of several tests for sweeping subckt M=
"Subcircuit multiliplier parameters (M) need to be handled
as an AST (enabling .STEP to be applied to subcircuit M)"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/492

This the beginnings of what will be a set of tests very similar to those
for issue 267 and issue 494.    In summary, the differences are:

1. ISSUE 267: use subcircuit multipliers in a bunch of scenarios,
   excluding setting them via .param or .global_param.

2. ISSUE 494: same as 267, but set subcircuit multipliers via
   .global_params.  However, do not modify the values of the
   .global_params (via .STEP or anything else) once simulation has
   started.

3. ISSUE 492: same as 494, but now sweep the subcircuit multipliers via
   .STEP, by sweeping the relevant .global_params and/or .params.

This commit only adds the test_a* series of tests.  This set of tests
were simple .DC tests, where the DC sweep was of a single point.  So,
instead of using .STEP I was able to use .DC to sweep the subcircuit
multiplier directly.

The other sets of tests (such as test_b* and test_c*, etc) are transient
tests, so I have to put a .STEP outer loop around the transient.  This
is fine, but it makes the required changes to the underlying scripts a
smidgen more complicated.  So those tests will be added in a future
commit.