Add a test for bugfix 35
Xyce/backlog/bugs#35 was about the "ceil" and "floor" functions
causing Xyce/ADMS to abort processing if they were found in an analog
function.

This test simply runs buildxyceplugin on a version of resistor.va that
has a useless analog function added to it.  That analog function uses
ceil and floor, and before bug 35 was fixed it would have caused
Xyce/ADMS to abort, saying ceil or floor (whichever it encountered
first) was invalid in analog functions.

The test passes if buildxyceplugin doesn't exit with an error, meaning
the plugin was created successfully and the bug is actually fixed.

Xyce/backlog/bugs#35