Fix incorrect code in shell scripts
All the shell scripts in this directory used a variable CIRCFILE to
try to archive a buildxyceplugin.log file when the test failed.  This
variable doesn't exist, it should have been CIRFILE.

That meant when trying to save "buildxyceplugin.log" under a unique
name so it doesn't get clobbered by subsequent tests, it was just
renaming it to "_buildxyceplugin.log" instead.  If more than one test
had failed, they'd all be clobbering each other anyway.

This has nothing to do with the bug that I'm working on on this
branch, except that I had copied one of these scripts and the test
failed, thereby not producing the expected uniquely named output file.