Add test for bugfix36
Using $strobe (or indeed any callfunction) in an analog function is
illegal, because Xyce/ADMS generates code for these callfunctions that
make reference to the instance pointer (so it can , for example, use
UserInfo(*this)) or to instance data (for something like $bound_step).

But rather than simply abort when a user tries to do this, it should
be possible to  ignore the callfunction and keep going.

This test is just a resistor with a completely pointless analog
function that calls $strobe.  The $strobe is ignored and the model
compiles anyway.  It didn't used to, it used to crash admsXml.

Xyce/backlog/bugs#36