Add a couple of tests of diode sidewall computation
These tests are of basic function of the new diode sidewall
computations in Xyce.

The gold standards are NOT Xyce output, they are the output of an
equivalent netlist from ngspice, massaged to look like Xyce output to
confuse xyce_verify.

diode_with_sidewall is a simple DC sweep that checks I-V and C-V
characteristics with three specific extra parameters specified.

diode_with_sidewall_temp does the same DC sweep, but with a
temperature sweep as well.  It, too, is compared to archived ngspice
output.

These two tests are necessary but insufficient tests of the new
contributions, as only a fraction of the new parameters are specified,
the rest being left at their defaults.

More are needed before declaring victory.

Xyce/backlog/project-backlog#157