Add test of diode breakdown with sidewall
I spent the late part of the day yesterday hunting down a disagreement
between Xyce with the new diode code and ngspice in the breakdown
region when "nbv" is anything but 1.0.  I got it nailed last night,
and this test verifies it.

A diode with sidewall contributions and the corrected handling of NBV
for all breakdown calculations is swept from -25 volts up to 1 volt.
Its diode current and capacitance is printed.

In this test, nbv is 0.33 (a randomly selected value that is very much
not 1) and BV is 15V, forcing the code to go through its BV/IBV
reconciliation process as well.

NBV is now correctly used both in the reconciliation and in the
breakdown region, and we match ngspice.

The gold standard associated with this test is not Xyce output, it is
output from ngspice on the equivalent netlist, massaged to look like
Xyce output so that xyce_verify.pl will treat it right.

Xyce/backlog/project-backlog#157