Add some basic EKV 2.6 tests
The NMOS and PMOS files simply take some somewhat arbitrarily
constructed, non-default EKV 2.6 model cards and run them through a
nested VD and VG voltage sweep to create IV curves.

The transient inverter netlist just takes an NMOS and PMOS ekv 2.6
device with completely default parameters (other than TNOM, which is
set to Xyce's default temperature instead of 25 degrees) and puts them
in inverter configuration.

The EKV model's defaults are clearly not set up in such a way that a
default NMOS and default PMOS are clearly matched, and the inverter
behaves strangely --- yes, it does seem to behave sort of like an
inverter, the on and off are very asymmetric and not at all what I'd
expect.

Yet this is what the model does.  Perhaps correctly extracted model
cards would work better.

The NMOS and PMOS sweeps are almost the same set of parameters with
only the VTO and TCV switched (to reflect NMOS vs. PMOS differences),
and they have extremely different IV curves.  But there it is --they
do in fact look like NMOS and PMOS IV curves.  EKV 2.6 comes with no
verification suite or even example model cards, and so it is basically
provided "as-is" with no way of knowing if you've got it all right.

These tests serve no purpose but as regression tests to assure there
is no accidental change in behavior somehow.

Xyce/backlog/project-backlog#449