Add tests of new integer variable code
Xyce/ADMS emitted incorrect code that would not compile if a user
assigned a probe-dependent expression to an integer variable in normal
context.

With the fix of this issue, the two verilog-a modules in this test now
compile.  If the bug were not fixed, they would both fail at the
compilation stage.

Xyce/backlog/project-backlog#106