Fix incorrect handling of error conditions
The .sh files for these 5 tests were copied from other VERILOGPLUGIN
tests that actually ran netlists using the verilog plugins produced.
But these tests don't --- they just check that the plugin compiles
properly.

When they failed (which they have not done since I made them) they
are supposed to the "buildxyceplugin.log" file to a file name that
includes the test name (so that the file is not clobbered by other
passing tests in the directory), but did so incorrectly by referencing
an undefined "CIRFILE" variable.

This commit assures that every one of these tests copies their
buildxyceplugin.log file to a uniquely named file so that the unlucky
soul who has to debug the problem can actually find the errors.