Fix perl syntax error
There were typos in these two shell scripts that only mattered when
the test was failing in some way, at which point they'd throw errors
from perl.

This showed up on my platform yesterday when there were other problems
with the code they are meant to test, and so the problematic lines of
perl got tripped.