Copy toy diode2.va model to test directory
I forgot, after committing a modified diode2.va to the Xyce repo, that
a second copy existed in the VERILOGPLUGIN directory that is actually
used in nightly testing.  This copies the modified version over.

The modification is to make the model initialize bias-independent
quantities in the @(initial_step) block instead of
@(initial_instance).  The former is a Verilog-A standard (which
Xyce/ADMS implements only half-correctly), and the latter a wholely
non-standard, ADMS-specific language extension.  Fortunately, the
"half-correct" part of the implementation is all we need for this.

Now we'll be making sure it continues to work.