Create a simple test for backlog issue 527
Having fixed Xyce/ADMS to allow a user to create analog functions that
take arguments of type different from the return type of the function,
we need to test a lot of special cases to make sure it's right.

This is the first test.

It is nothing but the same resistor that we have in the "toys"
directory, but with an added integer multiplicity parameter and all
the computation of conductance and resistor current moved down to an
analog function.  Since multiplicity is an integer and the current is
a real, the analog function is one that would have failed to build
prior to the fix.

Since it's not enough to show that the thing compiles, we must also
show that the code it produces actually runs and produces correct
results.  So we run a simple voltage divider where one resistor is a
built-in and the other a Verilog-A plugin with all the same
parameters (resistance 1K and multiplicity 2).  The resulting voltage
divider should have identical voltage drops across the two resistors
if the coding is correct.  It does, so it is.

This is only one of the special cases, because while the analog
function does have mixed argument types, the return type of both the
function and its additional output argument are reals, a use case that
we know to be correct already (modulo the issue of having ints in the
mix previously having been broken).  We will also need some tests
where the returned value is an int and which contains some int output
arguments.

Xyce/backlog/project-backlog#527