Add another test case for 527
Xyce/ADMS had serious problems processing analog functions whose
argument types were not all of the same type as the return type.

I have addressed the problems in the XML backend, and in a previous
test case I created a verilog-a model that made sure that a function
that returned a real but had one or more integer arguments worked
fine.

This test now checks that a function that takes a real argument but
returns an int works correctly when called from within another analog
function.

This is nowhere near enough to prove that all use cases are taken care
of, and I probably can't do that.  I will create one more test that
is a test of functions returning ints and taking reals, but called
from normal context instead.  It will be a minor variant of this one.

Xyce/backlog/project-backlog#527