Small fix to this test for RHEL7
This is a new test that I creaated recently for
PWL sources with .STEP.  In nightly testing it
failed on all the linux RHEL7 platforms.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/565
"Fix VPWL/IPWL independent sources to work with .STEP"

The reason it failed on RHEL7 is because the test was
relying on wildcards on the .PRINT line.  On RHEL7, the V(*)
wildcard resulted in the signals being output in different
order than they were output on OSX.

This commit removes the wildcard and enforces the
order of the outputs to specifically be:

V(7)  V(4)  V(5)  V(6)

which is the order that OSX produced.  This way I can
keep the gold standard untouched.  This seems to fix
the problem.