Adds tests for change to subcircuit processing in first pass of parser
Two tests have been added that explicit ensure that a subcircuit with
no name results in an error, which maintains the current behavior.  The
second test checks that an unused circuit with some basic formatting
issues will provide a warning but ultimately allow the simulator to
proceed.

The CircuitBlock_extractSubcircuitData_1.cir netlist had five subcircuits,
including one with no name.  After the change, this resulted in the
simulation exiting before emitting any other errors or warnings due to them
being deferred.  The unnamed circuit was removed from that test to illustrate
that subcircuits that were unused emitted warnings while subcircuits that
were used emitted errors.  The CircuitContext_beginSubcircuitContext_1.cir
netlist is a duplicate of the CircuitBlock_extractSubcircuitData_1.cir, so
it is not necessary and has been removed.

Xyce/backlog/project-backlog#555