Update tests for complex-valued .PRINT expressions
This work is related to my recent work on:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/28
"Make .PRINT expressions produce both real and imaginary part
automatically for frequency domain calculations"

My recent code fixes only affected these 8 tests.  They started failing
because they (1) had complex-valued expresssions on the .PRINT line and
(2) the gold standard only included the real part, as that is what Xyce
would output.

To fix this I had 2 choices.  I could update the tests to explicitly
request just the real part, and (mostly) leave the gold standards alone,
other than slightly changing a few of names in the headers.  Or, I could
just re-run the circuits and replace the gold standards, which would now
include the imaginary part of these expressions.  I chose the second
approach, partly to add tests for this issue.

The only potential glitch is that in at least one case the imaginary
parrt is near zero.  One of the reasons we sometimes use expressions on
the .print line is to offset a signal so that the waveform comparisons
don't produce false failures due to numerical noise.   For example, the
"invert_hb_lapack.cir" test case has this .print line:

.print hb {v(vout)+1.0} {v(in)+4.0}  {v(1)+4.0} ig(MP1) ig(MN1) is(MP1) id(MN1)  id(MP1)

The first 3 outputs: v(out), v(in) and v(1) are offset by 1.0 and 4.0,
for the reasons I describe.

As the offsets are real-valued, the imaginary part of these signals will
not be offset.   So, if there are test failures after these changes, I
predict it will be due to this lack of offset in the imaginary parts.