Fixes comment in test circuit
The comment in the test circuit indicated that it should not throw
an error, only give a warning.  Unfortunately, for subcircuits with
no name, it is not possible to continue on without a name.  The parser
hierarchy requires one for sanity.  This commit corrects the comment
on the expected behavior.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/555