Add complex-value .param and .print tests
This commit is related to two different issues:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/28
"Make .PRINT expressions produce both real and imaginary part
automatically for frequency domain calculations"

and also:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/29
"Allow .params to be complex-valued"

These are two very simple tests, which both contain a parameter that has
an imaginary value.  One test is an AC test and the other is a
DC test.  The circuits are identical, very simple RC circuits.  Also,
both print the imaginary parameter on the .PRINT line.

For the AC test, the .PRINT line is ".print ac v(1) {r0}", where r0 is the
imaginary parameter (.param r0={log(-1)}).   For the AC case, the
*.FD.prn file will automatically print both the real and imaginary parts
of {r0}.

For the DC case, the .PRINT line is
".print dc v(1) {r0} {re(r0)} {img(r0)}".

In the DC case, the outputter assumes that everything is real valued,
so for the {r0} output, it only outputs the real part of {r0}.  However,
even though DC is mainly a real-valued analysis, the imaginary part
of the expression for {r0} is still available, and can be obtained via
the output {img(r0)}.