Add extra code to shell script for valgrind
This commit is related to two different issues:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/28
"Make .PRINT expressions produce both real and imaginary part
automatically for frequency domain calculations"

and also:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/29
"Allow .params to be complex-valued"

I pushed these tests yesterday, and tagged them as valgrind
tests.  For the second test (test2.cir) this was fine,
as it fits the testing model that valgrind expects.

For the first test, however, that was not the case, as it is an
AC test that relies on a script.  So, last night it failed.
This test needed to extra code in the script to make it
work with valgrind.  This commit adds that code block.