Add HB test demonstrating complex param/.print
I previously created 4 tests which demonstrate that complex numbers and
expressions can be used in .param statements and on the .print line.
However, I had not included an HB test, and that was an oversight.

This commit adds a simple HB example.  It has 3 output files.  Two are
in the time domain and 1 in the frequency domain.  The complex outputs
are handled differently in each.

This commit is related to two different issues:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/28
"Make .PRINT expressions produce both real and imaginary part
automatically for frequency domain calculations"

and also:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/29
"Allow .params to be complex-valued"

This commit also includes a minor change to the comments in the
invert_hb.cir test case.  The comments in that file had indicated it was
an MPDE test, which it is not.