Test demonstrating efficiency in Bsrc tables
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/575
"Expression tables are inefficient when they are large and contain
expressions"

This test has a very large table in it, invoked via the expression
library in a Bsrc.  The table uses parameters (i.e. external
expressions) to set the values of the table.  A .STEP loop changes some
of the values.

Prior to recent changes in the expression library, this circuit took a
really long time to run, and all of the computational expense was in the
re-evaluation of the expressions in the table, even if they had not
changed.   The code was recently updated to be a lot smarter about this,
and now the circuit runs much faster.

As the resulting output file is huge, this test is *not* running
xyce_verify vs a gold standard.  The ONLY thing this test is really
doing is (1) making sure Xyce runs this without a fatal error and (2)
that it completes the run within a time limit.  I am setting the time
limit to 120 seconds.  That is much longer than it needs on my laptop,
but on the chance that some platforms need longer I've added some time.

As a test that runs in 120 seconds is a kind of long for nightly
regression (but only kind of) , I'm making this a weekly test.

As this test doesn't compare vs. a gold standard, there is no gold
standard in this commit. I originally created one, but it took too long
for xyce_verify to complete so I altered the test and removed it.