Another test for Bsrc tables
This is related to:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/575
"Expression tables are inefficient when they are large and contain
expressions"

This new test is very similar to the first tests I created in this
directory.  The main differences are:

(1) It is small enough to check via xyce_verify.
    This also means it is too small to demonstrate
    the performance improvements.

(2) The table expression has new syntax in it.
    Specifically, the parameters like VSS are
    surrounded by single quotes.  Prior to the
    work on this issue, the expression library
    would not parse nested single quotes.

Since it isn't practical to use xyce_verify on the main issue575 circuit,
this test provides at least one (very similar) circuit in which the
result is checked.  But this circuit by itself doesn't really prove
that issue 575 is fixed.  The weekly test, issue575.cir is required
for that.