Another test for Bsrc tables, updated
This is related to:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/575
"Expression tables are inefficient when they are large and contain
expressions"

The "small" test is intended to be run nightly.  But, I created the
issue575_small.cir.tags file by copying the original tags file, which
had just the weekly tag.  This commit changes the tag for the small test
from weekly to nightly.