Fix script errors that test valgrind_check result
I found a typo in a couple of test scripts. It’s the same typo in two
places (the original error was probably copied and pasted). I searched
for other occurrences and only turned up these two. The missing space
before the closing bracket caused incorrect behavior; the else branch
of the if block was always taken. This prevented the results from the
valgrind script from being reported, and just always reported success.
There was an error message about the bad syntax on the if line
(written to the stand error output file for the test) that was helpful
in tracking down this problem.

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/56