Modify test for device names w/weird characters
This test, "valid_chars_expressions.cir" was theoretically supposed to
test all the possible use cases for unusual characters in node names and
device names, when invoked from the expression library.  However, it
ignored use cases associated with {I(Vsrc)}.  (Vsrc currents used in
expressions, when the Vsrc has weird characters in the name).

This is related to this issue:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/595
"Expression library does not recognize device names that include square
brackets and a bunch of other unusual (but supported) characters"