Fix for this test, removing wildcard outputs
This test was intended to test "weird" characters in device names,
in expression library parsing.  So, the outputs on the .PRINT line
include a lot of things like {I(v<)}, etc.

One output was causing problems:  {I(v*)}.  Xyce allows device and node
names to include a star symbol as one of the characters.  However,
Xyce also supports the use of wildcards on the .PRINT line.

So, if you request I(v*) (non-expression at least), then Xyce will
intepret this as a wildcard specification, and will then print out *all*
the voltage source currents.

This wasn't what I intended to test, but it caused another problem.
For wildcards, the order they are printed is (currently at least)
compiler-dependent.  Different compilers sort the names differently,
so the outputs can be in a different order, and this messes up any gold
standards, which assume a fixed order.

So, the simplest fix is to remove the I(V*) ouptuts from this test.

As before, this is related to:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/595
"Expression library does not recognize device names that include square
brackets and a bunch of other unusual (but supported) characters"