Add test demonstrating local variation
"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

This new test (bsim3_test4) is a variation on a test that was already
in the ISSUE_239 directory, bsim3_test3.

The original test used global variation to apply a single distribution
function to a bunch of length parameters in 14 different BSIM3 mosfets.

This new test is the same circuit, but applies local variation.  As a
result, all 14 MOSFETs get a unique random number.