Add a test for issue 147
The commit adds a test to test the handling of the first failed
transient step out of the DC op. The circuit solves the DC op, but
failed the first step out of the DC op. Xyce previously reported
that it failed due to maximum number of local error tests which is
incorrect. It now reports correct error and time information

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/147