Fix check for warning text that no longer appears.
A change to the python interface that checks the number
of DAC's within a circuit is now working correctly.  But this
test would probe for the DAC's and expect a warning message
from Xyce there there are none in the circuit.  Now that the
python layer correctly sees that there are no DAC's it doesn't
call into Xyce for the names of the DAC's and Xyce doesn't
emit the warning.

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/59