Update test for local/global variation
"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

The tests in this directory were mostly to certify that ".option scale"
works correctly in Xyce.  However, many of them have sampling commands
and random expression operators.  They are all set up to compare the
results of two simulations; a circuit with "scale" vs an equivalent
circuit without it.

However, in one case (bsim4_test3) the two circuits were just different
enough for variation in one circuit to be treated as a global
variation, and local in the other one.  The reference circuit applied a
scalar manually, rather than using ".option scale", and this resulted in
an extra layer of indirection in the .param statements.

With the recent fixes to Xyce this caused that test to fail.  This
commit fixes the issue and makes the 2 circuits truly equivalent,
taking this into account.

This commit also adds the tag "expression" to various tag files in the
ISSUE_239 directory.