Update test print line
This test takes two different netlists and runs one in Xyce
using the -hspice-ext flag and the other through XDM translation.
The two netlists should be the same but one of them used a print
line like:

.PRINT TRAN V(X1.A2) V(X1.A1) V(X2.A1) V(X2.B) V(3)

While the second one has:

.PRINT TRAN V(X1.*) V(X2.A1) V(X2.B) V(3)

on some systems the second print would expand, in Xyce, to match
the first one but on other systems it would expand in a way
that reverses the first two columns as in “V(X1.*)” becomes
“V(X1.A1) V(X1.A2)”.  This would cause the comparison of the
output to fail because the columns are in the wrong order.

This has been fixed by having both netlists start with the SAME
print line.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/576