Update gold std to include -local_variation flag
"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

As part of my code changes for this issue, I added a command line option
to turn it off.  If Xyce is run with "-local_variation false", then
it will treat all random expression operators as global, and will
not treat any as local.  This is the previous behavior of Xyce, before
these changes.