Add a few more local variation tests
"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

This commit adds 3 new tests for local variation.

It also makes 2 of the existing tests (localVariation1 and
localVariation2) a little more robust.   All of these tests compare a
test circuit, which makes use of the implicit local variation
convention, to a reference circuit, which sets up the local variation
explicitly.

This works fine, except that the two circuits (test and reference) will
naturally have different names for the random parameters.  And, as these
names get sorted during setup, they could wind up in a different order.
So, these tests are now modified such that the random params will be
sorted in the same order.  Before, that was not the case.  This doesn't
matter for all of these, but it was easy to fix anyway.