Two more tests for local variation
"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

These involve subcircuits in a slightly more complex way than the
previous subcircuit tests.  They are more complex in that there are
multiple instances.

The two tests, 7 and 8, are both inverter chains.  The main difference
is that they are testing against different "reference" circuits.  Also,
the length of the chain is different, although that shouldn't matter
much.  (test 7 has 10 inverters, test 8 has 4).

The reference circuit for test 7 doesn't use subcircuits at all, and
applies the local variation in a manual brute force way.

The reference circuit for test 8 uses subcircuits, but still sets up the
local variation manually, and passes it in as a subcircuit instance
parameter argument.  This particular way of doing it didn't seem to work
correctly at first, but fortunatley that was caused by a netlist mistake
that I've managed to fix.

Also, it is worth noting that for these tests to work, it was crucial
that the list of uncertain parameters have equivalent orderings between
the test circuit and reference circuit.  Xyce sorts these parameters
internally, based on alphabetical order.