Taking this test out of parallel testing for now
This is one of several tests for this issue:

"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

Unfortunately, this test is failing in parallel.  The test circuit
itself is not failing, but the reference circuit is failing.

The reference circuit (being a reference circuit) doesn't exercise any
of the features that I added to support this issue.  So, it failing
does NOT indicate that the local variation fails in parallel.

I plan to write up the details explaining why this reference circuit
is failing in parallel in a new bug report.  Short summary: it has
nothing to do with local variation or UQ, but instead has to do
with how parameters (especially subcircuit instance parameters) are
passed around in parallel.  It is not a new issue at all; it has
in fact been in Xyce nearly forever.  But it is an obscure use case
so to the best of my knowledge it hasn't come up before.