Test for handling multiple options statements
This is the first of several tests for this issue:

"Modify Xyce parser so that it can handle multiple .options statements
for the same component (for example, support .options timeint
statements)"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/571

This particular test only tests out warning messages, for cases when the
user has specified not just the same ".options" line, but the same
parameter more than once.  In this test one redundancy is this one:

.options device voltlim=0
.options device voltlim=1

And the warning indicates that Xyce can only use one of these, and is
using the first one.

Also, this redundancy is present:

.options timeint reltol=1.0e-4 reltol=1.0e-5

And Xyce issues a warning that it cannot use both of these, and will use
the first one.

Otherwise, this is a simple transient circuit; this test doesn't bother
to check the calculation output.