Add a new test for issue 583
The new test is derived from a ring oscillator. It is equivalent to
the original netlist, but changed how the voltage sources are specified.
It is used to test the improvement on the source stepping. The circuit
would fail DC op because the source stepping did not work if both DC
and transient values are specified for a source before the work for
issue 583.  After the fix, the DC op can be solved using source stepping in Xyce

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/583