update PDE 2D BJT test
The PDE 2D BJT test specified conflicting DC and transient values
for a voltage source. The recent work changed which value to use
in this case for DC op calculation. This led to different result.
For the sources that have conflicting DC and transient values, the
behavior is not documented in Xyce and Xyce does not issue an
warning about which value to use. This will be fixed. Now
the test is updated to remove conflicting specifications in the voltage
source. This should fix the failure for the netlist