Add testing for source stepping with inconsistent DC and transient specification
The test specified an inconsistent DC and transient values for a
source. The source stepping did not work for this case before
the fix of issue 622. So the test was updated to use consistent
DC and transient specification. The issue 622 enables source
stepping to handle this case. The test is changed back to use
inconsistent values. It is used to test the source stepping

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/622