Adds 'noop' to a simple LTRA netlist to test time history fix
The time history vector, history size, and time index were only being
initialized during a DCOP calculation.  If a netlist skipped the DCOP
calculation by using UIC or NOOP, then the history vector would not be
initialized to a non-zero length.  Resulting calls to the LTRA devices
resulted in a seg fault during the access of the uninitialized time history
vector.  This has been fixed to initialize the time history vector and
related counters when either DCOP is being performed or the time history
is zero when there are LTRA devices in the circuit.  A customer circuit
illustrating this seg fault will now run to completion with these changes.

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/64