Update tests for local variaiton in parallel
This commit is related to these issues:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/631
"Expression based local variation doesn't always work with subcircuit
arguments"

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/609
"Global parameters that are applied to subcircuit parameters don't work
in parallel"

The 4 circuits that include the number 631 in the name are (obviously)
testing the first of the above issues.  Two of them use the
device-balance distribution tool and the other two use the default
first-come-first-serve distribution tool.

The second issue (609) is tested by the two tests, "localVariation8" and
"localVariation8_devbal".   They are identical, except that one uses the
default first-come-first-serve distribution tool, and the other  uses
the device balance distribution tool.