Update 4 tests for removed temp interpolation
These are 4 error message tests, which issued fatal errors for different
scenarios, in which the netlist incorrectly invoked temperature
interpolation.

They have been changed, to output a new error message, indicating the
temperature interpolation is no longer supported.

This is in support of this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/649
"Remove the temperature interpolation methods"