Update MOS1 temperature interpolation test
This is related to this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/649
"Remove the temperature interpolation methods"

The netlist for this test is not modified but I have changed the
test to be an error message test.  The netlist invokes the now-deleted
temperature interpolation feature of Xyce.   So, the test now ensures
that Xyce emits a coherent warning message and exits.