Update temp interpolation test to error test
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/649
"Remove the temperature interpolation methods"

The temperature interpolation feature of Xyce has been removed.  This
test (pjfet_temp) was only in the test suite to test this feature.
I have changed this test so that it is now an error test.  Xyce will
print a fatal error when this netlist is run.