Adds .STEP test of parameterized sources for sequential source stepping
This test identified that there was an issue with sequential source
stepping when a source was parametrized and swept with .STEP.  The
scaled parameters were not being reset between calls to sequential
source stepping.  This test is a small modification of the original
source_stepped_and_swept test that only runs the netlist with the
sweep and the individual netlists.  The concatenated prn file is
compared to the swept source prn file.  They should be exactly the
same.  They were not before the changes.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/692